updated server setup
couple odd things here that i changed:

1. we had split "server" and "worker" inits between the package init and
another worker module
2. kind of a massive anti pattern to have all of that work happening in
a package init file
3. moved app/server init logic into `server.py` to mirror `worker.py`

also linted the code moved into server.py and updated the import in
manage.py.
